♻️ Extract delegate swizzling to single implementation
