+++ Sources/AppcuesKit/Push/UNUserNotificationCenter+AutoConfig.swift
@@ -99,77 +99,23 @@ extension UNUserNotificationCenter {
             self.delegate = delegate
         }
 
-        swizzle(
-            delegate,
+        Swizzler.swizzle(
+            targetInstance: delegate,
             targetSelector: NSSelectorFromString("userNotificationCenter:didReceiveNotificationResponse:withCompletionHandler:"),
+            replacementOwner: UNUserNotificationCenter.self,
             placeholderSelector: #selector(appcues__placeholderUserNotificationCenterDidReceive),
             swizzleSelector: #selector(appcues__userNotificationCenterDidReceive)
         )
 
-        swizzle(
-            delegate,
+        Swizzler.swizzle(
+            targetInstance: delegate,
             targetSelector: NSSelectorFromString("userNotificationCenter:willPresentNotification:withCompletionHandler:"),
+            replacementOwner: UNUserNotificationCenter.self,
             placeholderSelector: #selector(appcues__placeholderUserNotificationCenterWillPresent),
             swizzleSelector: #selector(appcues__userNotificationCenterWillPresent)
         )
 
         return delegate
-    }
-
-    private func swizzle(
-        _ delegate: UNUserNotificationCenterDelegate,
-        targetSelector: Selector,
-        placeholderSelector: Selector,
-        swizzleSelector: Selector
-    ) {
-        // see if the currently assigned delegate has an implementation for the target selector already.
-        // these are optional methods in the protocol, and if they are not there already, we'll need to add
-        // a placeholder implementation so that we can consistently swap it with our override, which will attempt
-        // to call back into it, in case there was an implementation already - if we don't do this, we'll
-        // get invalid selector errors in these cases.
-        let originalMethod = class_getInstanceMethod(type(of: delegate), targetSelector)
-
-        if originalMethod == nil {
-            // this is the case where the existing delegate does not have an implementation for the target selector
-
-            guard let placeholderMethod = class_getInstanceMethod(UNUserNotificationCenter.self, placeholderSelector) else {
-                // this really shouldn't ever be nil, as that would mean the function defined a few lines below is no
-                // longer there, but we must nil check this call
-                return
-            }
-
-            // add the placeholder, so it can be swizzled uniformly
-            class_addMethod(
-                type(of: delegate),
-                targetSelector,
-                method_getImplementation(placeholderMethod),
-                method_getTypeEncoding(placeholderMethod)
-            )
-        }
-
-        // swizzle the new implementation to inject our own custom logic
-
-        // this should never be nil, as it would mean the function defined a few lines below is no longer there,
-        // but we must nil check this call.
-        guard let swizzleMethod = class_getInstanceMethod(UNUserNotificationCenter.self, swizzleSelector) else { return }
-
-        // add the swizzled version - this will only succeed once for this instance, if its already there, we've already
-        // swizzled, and we can exit early in the next guard
-        let addMethodResult = class_addMethod(
-            type(of: delegate),
-            swizzleSelector,
-            method_getImplementation(swizzleMethod),
-            method_getTypeEncoding(swizzleMethod)
-        )
-
-        guard addMethodResult,
-              let originalMethod = originalMethod ?? class_getInstanceMethod(type(of: delegate), targetSelector),
-              let swizzledMethod = class_getInstanceMethod(type(of: delegate), swizzleSelector) else {
-            return
-        }
-
-        // finally, here is where we swizzle in our custom implementation
-        method_exchangeImplementations(originalMethod, swizzledMethod)
     }
 
     @objc
